refactor: split MCP logic into mcp_components subfolder
